[Qt][WK2] Disable the flickable when pinching.
Reviewed by Kenneth Rohde Christiansen.

Set the interactivity of the internal flickable to false
when pinching so that it does not react to mouse events
which might break pinch zoom.
This is a temporal workaround and needed until the event
propagation in QtQuick is fixed.

* UIProcess/qt/QtViewportInteractionEngine.cpp:
(WebKit::QtViewportInteractionEngine::scaleAnimationStateChanged):
(WebKit::QtViewportInteractionEngine::pinchGestureStarted):
(WebKit::QtViewportInteractionEngine::pinchGestureEnded):

Identifier: 99336@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@111850 268f45cc-cd09-0410-ab3c-d52691b4dbfc