Fix some problems of text field decoration https://bugs.webkit.org/show_bug.cgi?id=82693
Reviewed by Dimitri Glazkov.

- Fix a problem that decorations are not removed when the input type is
changed.
- Add a comment to the 'willDetach' callback.

No new tests because the behavior changes are not visible yet.

* GNUMakefile.list.am: Add HTMLShadowElement.*.
* html/InputType.cpp:
(WebCore::InputType::destroyShadowSubtree):
Remove all of ShadowRoot contents, and add a <shadow> element to each of
them. We don't remove ShadowRoots from the tree because it's not
supported well.
* html/shadow/TextFieldDecorationElement.cpp:
(getDecorationRootAndDecoratedRoot): A helper function for decorate().
If the input element has a ShadowRoot with single <shadow> child, we
don't create new ShadowRoot and reuse it.
(WebCore::TextFieldDecorationElement::decorate):
Use getDecorationRootAndDecoratedRoot().
* html/shadow/TextFieldDecorationElement.h:
(TextFieldDecorator): Add a comment to willDetach().

Identifier: 100201@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@112830 268f45cc-cd09-0410-ab3c-d52691b4dbfc