Crash due to floating object lists not properly being cleared https://bugs.webkit.org/show_bug.cgi?id=74056
Patch by Ken Buchanan <kenrb@chromium.org> on 2012-04-02
Reviewed by David Hyatt.

Source/WebCore:

Add a check to clearFloats() that determines when intruding floats
are being cleared and not re-added. In this condition, ensure
children with floats are also getting layout because they might
need to have the same intruding floats cleared from their floating
object lists also.

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::clearFloats):

LayoutTests:

This test creates a condition where an intruding float is changed so
that it no longer intrudes. The child of a sibling was not getting
properly updated during the next layout.

* fast/block/float/intruding-float-not-removed-from-descendant-crash-expected.txt: Added
* fast/block/float/intruding-float-not-removed-from-descendant-crash.html: Added

Identifier: 100294@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@112935 268f45cc-cd09-0410-ab3c-d52691b4dbfc