Introduce another file-to-array script https://bugs.webkit.org/show_bug.cgi?id=83001
Reviewed by Kentaro Hara.

make-file-arrays.py is a script to convert a binary file to a C++
char array. The differences from css/make-css-file-arrays.pl and
inspector/xxd.pl are:

- No whitespace stripping for CSS files
- We can make conditional arrays.
e.g. An array enclosed with #if ENABLE(...) - #endif.

This script will be used to implement a calendar picker for <input
type=date>.

* make-file-arrays.py: Added.

Identifier: 100345@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@112988 268f45cc-cd09-0410-ab3c-d52691b4dbfc