[GTK] DRT missing didRunInsecureContent notification https://bugs.webkit.org/show_bug.cgi?id=59367
Patch by Simon Pena <spena@igalia.com> on 2012-04-03
Reviewed by Philippe Normand.

Source/WebKit/gtk:

Add a new signal "insecure-content-run" to the WebFrame to notify
when insecure HTTP content (such as CSS, an iframe or a script) is
run from a secure HTTPS WebFrame. Implement didRunInsecureContent
in gtk's FrameLoaderClient by means of emitting that signal.

* WebCoreSupport/FrameLoaderClientGtk.cpp: Implement
didRunInsecureContent by emitting WebFrame's
"insecure-content-run" signal
(WebKit::FrameLoaderClient::didRunInsecureContent):
* webkit/webkitwebframe.cpp: Add "insecure-content-run" signal
(webkit_web_frame_class_init):

Tools:

Connect DumpRenderTree to WebFrame's "insecure-content-run", in both
the main WebFrame or any other WebFrame created later. Added
"didRunInsecureContent" notification in the callback in order to
get the LayoutTests passing.

* DumpRenderTree/gtk/DumpRenderTree.cpp: Connect
to WebFrame's "insecure-content-run" signal and add
didRunInsecureContent notification in the callback to get the
LayoutTests passing.
(didRunInsecureContent):
(createWebView):

LayoutTests:

Unskip tests as new API is added to detect when a secure HTTPS page
loads content (such as CSS, images, iframes or a script) from
a non-secure HTTP origin.

* platform/gtk/Skipped: Unskip insecure-css-in-main-frame,
insecure-css-in-iframe, insecure-image-in-main-frame,
insecure-script-in-iframe, insecure-plugin-in-iframe,
redirect-http-to-https-script-in-iframe and
redirect-https-to-http-script-in-iframe since they pass after the
new API has been added.

Identifier: 100358@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@113001 268f45cc-cd09-0410-ab3c-d52691b4dbfc