Crash in SelectorChecker::checkOneSelector. https://bugs.webkit.org/show_bug.cgi?id=83040
Reviewed by Antti Koivisto.

Source/WebCore:

Test: fast/css/css-set-selector-text-crash.html

Removing the early bail when we detect that our selector text
hasn't changed, and we don't notify the styleSelectorChanged.
In fact, when we adopt the new selector list, the old one will
get destroyed and the styleSelectorChanged call needs to be made.

* css/CSSStyleRule.cpp:
(WebCore::CSSStyleRule::setSelectorText):

LayoutTests:

* fast/css/css-set-selector-text-crash-expected.txt: Added.
* fast/css/css-set-selector-text-crash.html: Added.


Identifier: 100447@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@113099 268f45cc-cd09-0410-ab3c-d52691b4dbfc