[GTK] ASSERT in SocketStreamHandleSoup.cpp https://bugs.webkit.org/show_bug.cgi?id=83123
Reviewed by Martin Robinson.

Do not try to reuse a GOwnPtr as calling outPtr() causes an
assertion if the pointer is already valid. Also do not try to
close the IOStream if it was not created.

This patch fixes
http/tests/websocket/tests/hybi/workers/worker-reload.html but I
am leaving it skipped until webkit.org/b/83124 is fixed.

* platform/network/soup/SocketStreamHandleSoup.cpp:
(WebCore::connectedCallback):

Identifier: 100530@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@113189 268f45cc-cd09-0410-ab3c-d52691b4dbfc