[CSS] Make makevalues.pl and makeprop.pl ignore '#'s. https://bugs.webkit.org/show_bug.cgi?id=83212
Reviewed by Tony Chang.

This patch is a side-effect of my intentions to make the
CMake-based ports work with GCC 4.7 without changing the
parameters currently passed to the preprocessor. "-P" is always
being passed to the preprocessor, but this breaks
dom/make_names.pl due to GCC not outputting empty lines with "-P"
anymore; if I switch to never passing "-P" (my aim), the CSS
scripts currently break because of the additional lines output by
the preprocessor. InFilesParser.pm and make-css-file-array.pl
already skip lines which start with '#', for example.

No new tests, tools plumbing.

* css/makeprop.pl:
* css/makevalues.pl:

Identifier: 100581@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@113244 268f45cc-cd09-0410-ab3c-d52691b4dbfc